set up apollo client provider
