+++ package.json
@@ -32,6 +32,7 @@
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
     "@fontsource/roboto": "^4.5.8",
+    "@material-ui/core": "^4.12.4",
     "@mui/icons-material": "^5.10.14",
     "@mui/material": "^5.10.14",
     "@types/react": "^18.0.25",
