+++ src/client/containers/Header.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 function Header(){
 
 return(
-    <div>Header</div>
+    <div>kafe logo (header)</div>
     );
 }
