+++ src/client/index.html
@@ -13,12 +13,11 @@
     />
     <meta name="format-detection" content="telephone=no" />
     <meta name="msapplication-tap-highlight" content="no" />
-    <!-- <link rel="icon" type="image/x-icon" href="../../../build/assets/favicon.ico"> -->
     <script defer="" src="/bundle.js"></script>
 </head>
 
   <body>
-    <div id="root"></div>
+    <div id="root" style="height:100vh"></div>
   </body>
 
 </html>
